+++ server/src/api/api.router.ts
@@ -1,19 +1,19 @@
 import express, { Request } from 'express';
 import { KibanaApi } from '../kibana/kibana-api';
-import orgRoutes from './org';
-import userRoutes from './user';
-import roleRoutes from './role';
-import rosterRoutes from './roster';
-import unitRoutes from './unit';
-import reingestRoutes from './reingest';
-import accessRequestRoutes from './access-request';
-import workspaceRoutes from './workspace';
-import notificationRoutes from './notification';
-import musterRoutes from './muster';
-import reportRoutes from './report-schema';
-import exportRoutes from './export';
-import orphanedRecordRoutes from './orphaned-record';
-import observationRoutes from './observation';
+import orgRoutes from './org/org.router';
+import userRoutes from './user/user.router';
+import roleRoutes from './role/role.router';
+import rosterRoutes from './roster/roster.router';
+import unitRoutes from './unit/unit.router';
+import reingestRoutes from './reingest/reingest.router';
+import accessRequestRoutes from './access-request/access-request.router';
+import workspaceRoutes from './workspace/workspace.router';
+import notificationRoutes from './notification/notification.router';
+import musterRoutes from './muster/muster.router';
+import reportRoutes from './report-schema/report-schema.router';
+import exportRoutes from './export/export.router';
+import orphanedRecordRoutes from './orphaned-record/orphaned-record.router';
+import observationRoutes from './observation/observation.router';
 import { User } from './user/user.model';
 import { Org } from './org/org.model';
 import { UserRole } from './user/user-role.model';
