+++ server/src/api/export/export.controller.ts
@@ -15,7 +15,7 @@ import {
   ApiRequest,
   OrgParam,
   PaginatedQuery,
-} from '../index';
+} from '../api.router';
 import {
   RosterColumnInfo,
   RosterColumnValue,
