+++ server/src/api/org/org.router.ts
@@ -5,7 +5,7 @@ import {
   requireOrgAccess,
   requireRolePermission,
   requireRootAdmin,
-} from '../../auth';
+} from '../../auth/auth-middleware';
 
 const router = express.Router() as any;
