+++ server/src/api/roster/roster.router.ts
@@ -3,7 +3,7 @@ import multer from 'multer';
 import path from 'path';
 import bodyParser from 'body-parser';
 import controller from './roster.controller';
-import { requireInternalUser, requireOrgAccess, requireRolePermission } from '../../auth';
+import { requireInternalUser, requireOrgAccess, requireRolePermission } from '../../auth/auth-middleware';
 
 const rosterUpload = multer({
   storage: multer.diskStorage({
